requirements.txt
setuptools
newspaper4k